[CfgEditor] Integrate multiple js into one
There were multiple js files in `media` for maintenance.
However, it causes unintended error in remote-ssh environment.
To resolve this, and for better performance,
multiple js files should be integrated as one.

ONE-vscode-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>
